sua gio hang

+++ LaptopStore/LaptopStore/Areas/Admin/Controllers/HomeController.cs
@@ -23,11 +23,11 @@ namespace LaptopStore.Areas.Admin.Controllers
                 ViewBag.moneyInMoth = 0;
                 ViewBag.totalMoney = 0;
             }
-            int[] dataOfYear = { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 };
+            decimal[] dataOfYear = { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 };
             for (int i =0; i<12; i++)
             {
                 int month = i + 1;
-                int data = 0;
+                decimal data = 0;
                 try
                 {
                    data = db.bills.Where(b => b.dateCreate.Month == month && b.status == "Đã giao").Sum(b => b.total);
